Moving common data structure to a shared structure
This is the first step to make sure that a more component based approached
is used throughout the source code. the structs sub package will contain
all data structures that are resused between different packages.

Notable in this refactoring is a generic Grid struct that is currently
used for collision detection, but can be resused for map generation, path
planning etc.